[FLINK-6196] [table] Move the methods to some util classes
