+++ flink-libraries/flink-table/src/main/scala/org/apache/flink/table/functions/TableFunction.scala
@@ -18,13 +18,11 @@
 
 package org.apache.flink.table.functions
 
-import org.apache.calcite.rex.{RexLiteral, RexNode}
-import org.apache.calcite.sql.`type`.SqlTypeName
 import org.apache.flink.api.common.typeinfo.TypeInformation
-import org.apache.flink.table.expressions.{Expression, Literal, TableFunctionCall}
+import org.apache.flink.table.expressions.{Expression, TableFunctionCall}
+import org.apache.flink.table.functions.utils.UserDefinedFunctionUtils
 import org.apache.flink.util.Collector
 
-import scala.collection.JavaConversions._
 
 /**
   * Base class for a user-defined table function (UDTF). A user-defined table functions works on
@@ -90,24 +88,8 @@ abstract class TableFunction[T] extends UserDefinedFunction {
     * @return [[Expression]] in form of a [[TableFunctionCall]]
     */
   final def apply(params: Expression*)(implicit typeInfo: TypeInformation[T]): Expression = {
-    buildTableFunctionCall(getClass.getSimpleName, typeInfo, params: _*)
-  }
-
-  /**
-    * Build a TableFunctionCall, a name and a sequence of params will determine a unique
-    * [[TableFunctionCall]]
-    *
-    * @param name function name
-    * @param implicitResultType If no result type returned, it will use this type.
-    * @param params The input expressions
-    * @return A unique [[TableFunctionCall]]
-    */
-  private[table] def buildTableFunctionCall(name: String,
-                                            implicitResultType: TypeInformation[_],
-                                            params: Expression*): TableFunctionCall = {
-    val arguments = expressionsToArguments(params: _*)
-    val resultType = getResultType(arguments, implicitResultType)
-    TableFunctionCall(name, this, params, resultType)
+    UserDefinedFunctionUtils.buildTableFunctionCall(
+      getClass.getSimpleName, this, typeInfo, params: _*)
   }
 
   /**
@@ -127,43 +109,6 @@ abstract class TableFunction[T] extends UserDefinedFunction {
     }
   }
 
-  /**
-    * Transform the expressions or rex nodes to Objects
-    * Only literal expressions will be passed, nulls for non-literal ones
-    *
-    * @param params actual parameters of the function
-    * @return A java List of the Objects
-    */
-  private[table] def expressionsToArguments(params: Expression*): java.util.List[AnyRef] = {
-    params.map {
-      case exp: Literal =>
-        exp.value.asInstanceOf[AnyRef]
-      case _ =>
-        null
-    }
-  }
-
-  private[table] def rexNodesToArguments(
-          rexNodes: java.util.List[RexNode]): java.util.List[AnyRef] = {
-    rexNodes.map {
-      case rexNode: RexLiteral =>
-        val value = rexNode.getValue2
-        rexNode.getType.getSqlTypeName match {
-          case SqlTypeName.INTEGER =>
-            value.asInstanceOf[Long].toInt.asInstanceOf[AnyRef]
-          case SqlTypeName.SMALLINT =>
-            value.asInstanceOf[Long].toShort.asInstanceOf[AnyRef]
-          case SqlTypeName.TIMESTAMP =>
-            value.asInstanceOf[Long].toByte.asInstanceOf[AnyRef]
-          case SqlTypeName.FLOAT =>
-            value.asInstanceOf[Double].toFloat.asInstanceOf[AnyRef]
-          case _ =>
-            value.asInstanceOf[AnyRef]
-        }
-      case _ =>
-        null
-    }
-  }
 
   override def toString: String = getClass.getCanonicalName
